backend/requirements.txt
fastapi==0.127.1
uvicorn==0.24.0
python-multipart==0.0.6
email-validator==2.3.0
sqlmodel==0.0.30
SQLAlchemy==2.0.45
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
python-jose==3.3.0
cryptography==46.0.3
passlib==1.7.4
bcrypt==4.0.1
slowapi==0.1.9
limits==5.6.0